LayoutTests/inspector/canvas/css-canvas-clients-expected.txt

Test that CanvasAgent tracks changes in the client nodes of a CSS canvas.


== Running test suite: Canvas.CSSCanvasClients
-- Running test case: Canvas.CSSCanvasClients.InitialLoad
PASS: CanvasManager should have one canvas.
PASS: Canvas should have CSS name "css-canvas".
PASS: There should be no client nodes.

-- Running test case: Canvas.CSSCanvasClients.Create
PASS: Canvas with created client should have CSS name "css-canvas".
PASS: There should be one client node.
PASS: Client node "div" is valid.

-- Running test case: Canvas.CSSCanvasClients.Destroy
PASS: Canvas with destroyed client should have CSS name "css-canvas".
PASS: There should be no client nodes.

-- Running test case: Canvas.CSSCanvasClients.InvalidCanvasId
PASS: Should produce an error.
Error: Missing canvas for given canvasId
